# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Call Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /phone-numbers:
    get:
      operationId: listPhoneNumbers
      tags: [phone-numbers]
      summary: List all provisioned phone numbers
      responses:
        "200":
          description: List of phone numbers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PhoneNumber"
    post:
      operationId: provisionPhoneNumber
      tags: [phone-numbers]
      summary: Provision a new phone number from Twilio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneNumberInput"
      responses:
        "201":
          description: Provisioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumber"
  /phone-numbers/search:
    get:
      operationId: searchAvailableNumbers
      tags: [phone-numbers]
      summary: Search available Twilio phone numbers
      parameters:
        - name: areaCode
          in: query
          schema:
            type: ["string", "null"]
        - name: contains
          in: query
          schema:
            type: ["string", "null"]
        - name: country
          in: query
          schema:
            type: string
            default: US
        - name: tollFree
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Available numbers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailableNumber"
  /phone-numbers/{id}:
    get:
      operationId: getPhoneNumber
      tags: [phone-numbers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumber"
    patch:
      operationId: updatePhoneNumber
      tags: [phone-numbers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneNumberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumber"
    delete:
      operationId: releasePhoneNumber
      tags: [phone-numbers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Released
  /phone-numbers/{id}/test-call:
    post:
      operationId: testCall
      tags: [phone-numbers]
      summary: Place a test call to verify forwarding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestCallInput"
      responses:
        "200":
          description: Call initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCallResult"
...
components:
  schemas:
...
    PhoneNumber:
      type: object
      required: [id, number, friendlyName, callerIdName, forwardTo, ringCount, answerMode, isActive, createdAt]
      properties:
        id:
          type: integer
        number:
          type: string
        twilioSid:
          type: ["string", "null"]
        friendlyName:
          type: string
        callerIdName:
          type: string
        companyId:
          type: ["integer", "null"]
        forwardTo:
          type: ["string", "null"]
        ringCount:
          type: integer
        answerMode:
          type: string
          enum: [forward, ai_voice, voicemail, reject]
        aiSystemPrompt:
          type: ["string", "null"]
        voicemailGreeting:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    PhoneNumberInput:
      type: object
      required: [number, friendlyName, callerIdName, answerMode]
      properties:
        number:
          type: string
        twilioSid:
          type: ["string", "null"]
        friendlyName:
          type: string
        callerIdName:
          type: string
        companyId:
          type: ["integer", "null"]
        forwardTo:
          type: ["string", "null"]
        ringCount:
          type: integer
          default: 4
        answerMode:
          type: string
          enum: [forward, ai_voice, voicemail, reject]
        aiSystemPrompt:
          type: ["string", "null"]
        voicemailGreeting:
          type: ["string", "null"]
    PhoneNumberUpdate:
      type: object
      properties:
        friendlyName:
          type: ["string", "null"]
        callerIdName:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        forwardTo:
          type: ["string", "null"]
        ringCount:
          type: ["integer", "null"]
        answerMode:
          type: ["string", "null"]
          enum: [forward, ai_voice, voicemail, reject, null]
        aiSystemPrompt:
          type: ["string", "null"]
        voicemailGreeting:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    AvailableNumber:
      type: object
      required: [phoneNumber, friendlyName, locality, region, isoCountry]
      properties:
        phoneNumber:
          type: string
        friendlyName:
          type: string
        locality:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        isoCountry:
          type: string
        capabilities:
          type: object
          properties:
            voice:
              type: boolean
            sms:
              type: boolean
            mms:
              type: boolean
    TestCallInput:
      type: object
      required: [toNumber]
      properties:
        toNumber:
          type: string
    TestCallResult:
      type: object
      required: [callSid, status]
      properties:
        callSid:
          type: string
        status:
          type: string
...