# lib/api-zod/src/generated/types/availableNumber.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Call Center API
 * OpenAPI spec version: 0.1.0
 */
import type { AvailableNumberCapabilities } from './availableNumberCapabilities';

export interface AvailableNumber {
  phoneNumber: string;
  friendlyName: string;
  /** @nullable */
  locality: string | null;
  /** @nullable */
  region: string | null;
  /** @nullable */
  postalCode?: string | null;
  isoCountry: string;
  capabilities?: AvailableNumberCapabilities;
}