# lib/api-zod/src/generated/types/phoneNumber.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Call Center API
 * OpenAPI spec version: 0.1.0
 */
import type { PhoneNumberAnswerMode } from './phoneNumberAnswerMode';

export interface PhoneNumber {
  id: number;
  number: string;
  /** @nullable */
  twilioSid?: string | null;
  friendlyName: string;
  callerIdName: string;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  forwardTo: string | null;
  ringCount: number;
  answerMode: PhoneNumberAnswerMode;
  /** @nullable */
  aiSystemPrompt?: string | null;
  /** @nullable */
  voicemailGreeting?: string | null;
  isActive: boolean;
  createdAt: string;
}