# lib/api-zod/src/generated/types/phoneNumberUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Call Center API
 * OpenAPI spec version: 0.1.0
 */
import type { PhoneNumberUpdateAnswerMode } from './phoneNumberUpdateAnswerMode';

export interface PhoneNumberUpdate {
  /** @nullable */
  friendlyName?: string | null;
  /** @nullable */
  callerIdName?: string | null;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  forwardTo?: string | null;
  /** @nullable */
  ringCount?: number | null;
  /** @nullable */
  answerMode?: PhoneNumberUpdateAnswerMode;
  /** @nullable */
  aiSystemPrompt?: string | null;
  /** @nullable */
  voicemailGreeting?: string | null;
  /** @nullable */
  isActive?: boolean | null;
}